+++ src/views/NotFoundView.vue
@@ -1,0 +1,8 @@
+<template>
+  <section class="vc-not-found">
+    <p>{{ $t('core.notFound.page') }}</p>
+    <p>
+      <router-link to="/">{{ $t('core.nav.home') }}</router-link>
+    </p>
+  </section>
+</template>